# Conflicts:
#	src/LGML/Source/QueuedNotifier.h
#	src/LGML/Source/TimeManager.cpp

+++ src/LGML/Source/NodeBase.cpp
@@ -327,7 +327,7 @@ void NodeBase::updateRMS(const AudioBuffer<float>& buffer, float &targetRmsValue
 
 
         targetRMSChannelValues.set(i, (s>targetRMSChannelValues.getUnchecked(i))?s:
-                                        s>lowThresh?targetRMSChannelValues.getUnchecked(i)*decayFactor:
+                                        s>lowThresh?targetRMSChannelValues.getUnchecked(i)*(float)decayFactor:
                                         0);
 		globalS = jmax(s, globalS);
 	}
